backend-django/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=10.0
tensorflow>=2.15
numpy>=1.24
